+++ jwst/stpipe/pipeline.py
@@ -37,6 +37,8 @@ from . import config_parser
 from . import Step
 from . import crds_client
 from . import log
+from ..lib.class_property import ClassInstanceMethod
+
 
 class Pipeline(Step):
     """
@@ -187,7 +189,8 @@ class Pipeline(Step):
         try:
             ref_file = crds_client.get_reference_file(dataset,
                                                       pars_model.meta.reftype,
-                                                      observatory=observatory)
+                                                      observatory=observatory,
+                                                      asn_exptypes=['science'])
         except (AttributeError, exceptions.CrdsError, exceptions.CrdsLookupError):
             log.log.debug(f'{pars_model.meta.reftype.upper()}: No parameters found')
         else:
@@ -321,3 +324,35 @@ class Pipeline(Step):
             return False
         else:
             return True
+
+    @ClassInstanceMethod
+    def get_pars(pipeline, full_spec=True):
+        """Retrieve the configuration parameters of a pipeline
+
+        The pipeline, and all referenced substeps, parameters
+        are retrieved.
+
+        Parameters
+        ----------
+        step : `Pipeline`-derived class or instance
+
+        full_spec : bool
+            Return all parameters, including parent-specified parameters.
+            If `False`, return only parameters specific to the class/instance.
+
+        Returns
+        -------
+        pars : dict
+            Keys are the parameters and values are the values.
+        """
+        pars = super().get_pars(full_spec=full_spec)
+        for step_name, step_class in pipeline.step_defs.items():
+
+            # If a step has already been instantiated, get its parameters
+            # from the instantiation. Otherwise, retrieve from the class
+            # itself.
+            try:
+                pars[step_name] = getattr(pipeline, step_name).get_pars(full_spec=full_spec)
+            except AttributeError:
+                pars[step_name] = step_class.get_pars(full_spec=full_spec)
+        return pars