+++ jwst/stpipe/tests/steps/__init__.py
@@ -178,6 +178,18 @@ class StepWithModel(Step):
         return model
 
 
+class EmptyPipeline(Pipeline):
+    """A pipeline that has no substeps"""
+
+    spec = """
+    par1 = string(default='Name the atomizer') # Control the frobulization
+    """
+
+    def process(self, *args):
+
+        return args
+
+
 class ProperPipeline(Pipeline):
     """Pipeline with proper output setup"""
 
@@ -226,10 +238,12 @@ class SavePipeline(Pipeline):
         return r
 
 
-class FooPipeline(Pipeline):
+class MakeListPipeline(Pipeline):
     """A pipeline that calls MakeListStep"""
 
-    spec = ""
+    spec = """
+    par1 = string(default='Name the atomizer') # Control the frobulization
+    """
 
     step_defs = {
         'make_list': MakeListStep,