+++ jwst/stpipe/tests/test_step.py
@@ -13,7 +13,7 @@ from jwst.stpipe.config_parser import ValidationError
 from jwst.stpipe import crds_client
 from jwst.extern.configobj.configobj import ConfigObj
 
-from .steps import MakeListStep
+from .steps import EmptyPipeline, MakeListPipeline, MakeListStep
 from .util import t_path
 
 from crds.core.exceptions import CrdsLookupError
@@ -32,7 +32,11 @@ REFPIXSTEP_CRDS_MIRI_PARS = {
     'use_side_ref_pixels': False
 }
 
-@pytest.mark.xfail(reason="To be handled under a different ticket")
+
+@pytest.mark.xfail(
+    reason='Need to make regression with actual CRDS data',
+    run=False,
+)
 def test_parameters_from_crds():
     """Test retrieval of parameters from CRDS"""
     step_class = REFPIXSTEP_CRDS_MIRI_PARS['class']
@@ -40,12 +44,12 @@ def test_parameters_from_crds():
     pars = step_class.get_config_from_reference(data)
     assert pars == REFPIXSTEP_CRDS_MIRI_PARS
 
-@pytest.mark.xfail(reason="To be handled under a different ticket")
+
 def test_parameters_from_crds_fail():
     """Test retrieval of parameters from CRDS"""
     data = datamodels.open(t_path(join('data', 'miri_data.fits')))
     data.meta.instrument.name = 'NIRSPEC'
-    pars = crds_client.get_parameters_from_reference(RefPixStep, data)
+    pars = RefPixStep.get_config_from_reference(data)
     assert not len(pars)
 
 
@@ -83,50 +87,290 @@ def test_saving_pars(tmpdir):
 
 
 @pytest.mark.parametrize(
-    'step_obj, expected',
+    'step_obj, full_spec, expected',
     [
-        (MakeListStep,
-         datamodels.StepParsModel(
-             {'parameters': {
-                 'par1': 'float() # Control the frobulization',
-                 'par2': 'string() # Reticulate the splines',
-                 'par3': False,
-                 'class': 'jwst.stpipe.tests.steps.MakeListStep',
-                 'name': 'MakeListStep',
-             }}
-         )
+        # #############################################
+        # Test `get_pars_model` with `full_spec = True`
+        # #############################################
+        #
+        # Step Class with mix of required and optional parameters
+        (MakeListStep, True,
+         {'class': 'jwst.stpipe.tests.steps.MakeListStep',
+          'name': 'MakeListStep',
+          'pre_hooks': [],
+          'post_hooks': [],
+          'output_file': None,
+          'output_dir': None,
+          'output_ext': '.fits',
+          'output_use_model': False,
+          'output_use_index': True,
+          'save_results': False,
+          'skip': False,
+          'suffix': None,
+          'search_output_file': True,
+          'input_dir': None,
+          'par3': False,
+          'par1': 'float() # Control the frobulization',
+          'par2': 'string() # Reticulate the splines'}
         ),
-        (MakeListStep(par1=0., par2='from args'),
-         datamodels.StepParsModel({'parameters': {
-             'par1': 0., 'par2': 'from args', 'par3': False,
-             'class': 'jwst.stpipe.tests.steps.MakeListStep',
-             'name': 'MakeListStep'
-         }})
+        # Instance with parameters set
+        (MakeListStep(par1=0., par2='from args'), True,
+         {'class': 'jwst.stpipe.tests.steps.MakeListStep',
+          'name': 'MakeListStep',
+          'pre_hooks': [],
+          'post_hooks': [],
+          'output_file': None,
+          'output_dir': None,
+          'output_ext': '.fits',
+          'output_use_model': False,
+          'output_use_index': True,
+          'save_results': False,
+          'skip': False,
+          'suffix': None,
+          'search_output_file': True,
+          'input_dir': '',
+          'par1': 0.0,
+          'par2': 'from args',
+          'par3': False}
         ),
-        (Step.from_config_file(t_path(join('steps', 'jwst_generic_pars-makeliststep_0002.asdf'))),
-         ParsModelWithPar3
-        )
+        # ##############################################
+        # Test `get_pars_model` with `full_spec = False`
+        # ##############################################
+        #
+        # Step Class with mix of required and optional parameters
+        (MakeListStep, False,
+         {'class': 'jwst.stpipe.tests.steps.MakeListStep',
+          'name': 'MakeListStep',
+          'par3': False,
+          'par1': 'float() # Control the frobulization',
+          'par2': 'string() # Reticulate the splines'}
+        ),
+        # Instance with parameters set
+        (MakeListStep(par1=0., par2='from args'), False,
+         {'class': 'jwst.stpipe.tests.steps.MakeListStep',
+          'name': 'MakeListStep',
+          'par1': 0.0,
+          'par2': 'from args',
+          'par3': False}
+        ),
     ]
 )
-def test_getpars_model(step_obj, expected):
+def test_getpars_model(step_obj, full_spec, expected):
     """Test retreiving of configuration parameters"""
-    assert step_obj.get_pars_model().parameters == expected.parameters
+    assert step_obj.get_pars_model(full_spec=full_spec).parameters.instance == expected
 
 
 @pytest.mark.parametrize(
-    'step_obj, expected',
+    'step_obj, full_spec, expected',
     [
-        (MakeListStep, {
+        # #######################################
+        # Test `get_pars` with `full_spec = True`
+        # #######################################
+        #
+        # Step Class with mix of required and optional parameters
+        (MakeListStep, True, {
+            'pre_hooks': [],
+            'post_hooks': [],
+            'output_file': None,
+            'output_dir': None,
+            'output_ext': '.fits',
+            'output_use_model': False,
+            'output_use_index': True,
+            'save_results': False,
+            'skip': False,
+            'suffix': None,
+            'search_output_file': True,
+            'input_dir': None,
+            'par3': False,
             'par1': 'float() # Control the frobulization',
-            'par2': 'string() # Reticulate the splines',
+            'par2': 'string() # Reticulate the splines'
+        }),
+        # Instance with parameters set
+        (MakeListStep(par1=0., par2='from args'), True, {
+            'pre_hooks': [],
+            'post_hooks': [],
+            'output_file': None,
+            'output_dir': None,
+            'output_ext': '.fits',
+            'output_use_model': False,
+            'output_use_index': True,
+            'save_results': False,
+            'skip': False,
+            'suffix': None,
+            'search_output_file': True,
+            'input_dir': '',
+            'par1': 0.0,
+            'par2': 'from args',
             'par3': False
         }),
-        (MakeListStep(par1=0., par2='from args'), {'par1': 0., 'par2': 'from args', 'par3': False}),
+        #
+        # Pipeline Class with sub-step with mix of required and optional parameters
+        #
+        (MakeListPipeline, True, {
+            'pre_hooks': [],
+            'post_hooks': [],
+            'output_file': None,
+            'output_dir': None,
+            'output_ext': '.fits',
+            'output_use_model': False,
+            'output_use_index': True,
+            'save_results': False,
+            'skip': False,
+            'suffix': None,
+            'search_output_file': True,
+            'input_dir': None,
+            'par1': 'Name the atomizer',
+            'make_list': {
+                'pre_hooks': [],
+                'post_hooks': [],
+                'output_file': None,
+                'output_dir': None,
+                'output_ext': '.fits',
+                'output_use_model': False,
+                'output_use_index': True,
+                'save_results': False,
+                'skip': False,
+                'suffix': None,
+                'search_output_file': True,
+                'input_dir': None,
+                'par3': False,
+                'par1': 'float() # Control the frobulization',
+                'par2': 'string() # Reticulate the splines'
+            }
+        }),
+        #
+        # Pipeline Instance with all parameters set
+        #
+        (MakeListPipeline(
+            par1='Instantiated', steps={'make_list': {'par1': 0., 'par2': 'sub-instantiated'}}
+        ), True, {
+            'pre_hooks': [],
+            'post_hooks': [],
+            'output_file': None,
+            'output_dir': None,
+            'output_ext': '.fits',
+            'output_use_model': False,
+            'output_use_index': True,
+            'save_results': False,
+            'skip': False,
+            'suffix': None,
+            'search_output_file': True,
+            'input_dir': '',
+            'par1': 'Instantiated',
+            'make_list': {
+                'pre_hooks': [],
+                'post_hooks': [],
+                'output_file': None,
+                'output_dir': None,
+                'output_ext': '.fits',
+                'output_use_model': False,
+                'output_use_index': True,
+                'save_results': False,
+                'skip': False,
+                'suffix': None,
+                'search_output_file': True,
+                'input_dir': '',
+                'par1': 0.0,
+                'par2': 'sub-instantiated',
+                'par3': False
+            }
+        }),
+        #
+        # Pipeline class without any sub-steps
+        #
+        (EmptyPipeline, True, {
+            'pre_hooks': [],
+            'post_hooks': [],
+            'output_file': None,
+            'output_dir': None,
+            'output_ext': '.fits',
+            'output_use_model': False,
+            'output_use_index': True,
+            'save_results': False,
+            'skip': False,
+            'suffix': None,
+            'search_output_file': True,
+            'input_dir': None,
+            'par1': 'Name the atomizer'
+        }),
+        #
+        # Pipeline instance without any sub-steps
+        #
+        (EmptyPipeline(par1='Instantiated'), True, {
+            'pre_hooks': [],
+            'post_hooks': [],
+            'output_file': None,
+            'output_dir': None,
+            'output_ext': '.fits',
+            'output_use_model': False,
+            'output_use_index': True,
+            'save_results': False,
+            'skip': False,
+            'suffix': None,
+            'search_output_file': True,
+            'input_dir': '',
+            'par1': 'Instantiated'
+        }),
+        # ######################################
+        # Test `get_pars` with `full_spec=False`
+        # ######################################
+        #
+        # Step Class with mix of required and optional parameters
+        #
+        (MakeListStep, False, {
+            'par3': False,
+            'par1': 'float() # Control the frobulization',
+            'par2': 'string() # Reticulate the splines'
+        }),
+        #
+        # Instance with parameters set
+        #
+        (MakeListStep(par1=0., par2='from args'), False, {
+            'par1': 0.0,
+            'par2': 'from args',
+            'par3': False
+        }),
+        #
+        # Pipeline Class with sub-step with mix of required and optional parameters
+        #
+        (MakeListPipeline, False, {
+            'par1': 'Name the atomizer',
+            'make_list': {
+                'par3': False,
+                'par1': 'float() # Control the frobulization',
+                'par2': 'string() # Reticulate the splines'
+            }
+        }),
+        #
+        # Pipeline Instance with all parameters set
+        #
+        (MakeListPipeline(
+            par1='Instantiated', steps={'make_list': {'par1': 0., 'par2': 'sub-instantiated'}}
+        ), False, {
+            'par1': 'Instantiated',
+            'make_list': {
+                'par1': 0.0,
+                'par2': 'sub-instantiated',
+                'par3': False
+            }
+        }),
+        #
+        # Pipeline class without any sub-steps
+        #
+        (EmptyPipeline, False, {
+            'par1': 'Name the atomizer'
+        }),
+        #
+        # Pipeline instance without any sub-steps
+        #
+        (EmptyPipeline(par1='Instantiated'), False, {
+            'par1': 'Instantiated'
+        }),
     ]
 )
-def test_getpars(step_obj, expected):
+def test_getpars(step_obj, full_spec, expected):
     """Test retreiving of configuration parameters"""
-    assert step_obj.get_pars() == expected
+    assert step_obj.get_pars(full_spec=full_spec) == expected
 
 
 def test_hook():
@@ -183,7 +427,6 @@ def test_step_from_python():
     assert result == 3
 
 
-@pytest.mark.xfail(reason="To be handled under a different ticket")
 def test_step_from_python_simple():
     from .steps import AnotherDummyStep
 
@@ -192,7 +435,6 @@ def test_step_from_python_simple():
     assert result == 3
 
 
-@pytest.mark.xfail(reason="To be handled under a different ticket")
 def test_step_from_python_simple2():
     from .steps import AnotherDummyStep
 
@@ -365,7 +607,7 @@ def test_step_from_commandline_par_precedence(command_line_pars, command_line_co
 
         reference_file_map[reference_type] = str(reference_path)
 
-    def mock_get_reference_file(dataset, reference_file_type, observatory=None):
+    def mock_get_reference_file(dataset, reference_file_type, observatory=None, asn_exptypes=None):
         if reference_file_type in reference_file_map:
             return reference_file_map[reference_file_type]
         else: