+++ jwst/tests_nightly/general/niriss/test_nis_wfss_spec2.py
@@ -1,7 +1,7 @@
 import pytest
 
 from jwst.pipeline.calwebb_spec2 import Spec2Pipeline
-
+from jwst.pipeline.collect_pipeline_cfgs import collect_pipeline_cfgs
 from jwst.tests.base_classes import BaseJWSTTest, raw_from_asn
 
 
@@ -15,18 +15,17 @@ class TestSpec2Pipeline(BaseJWSTTest):
         """
         Regression test of calwebb_spec2 pipeline performed on NIRISS WFSS data.
         """
+        # Collect data
         asn_file = self.get_data(self.test_dir,
                                  'jw87600-a3001_20171109T145456_spec2_001_asn.json')
         for file in raw_from_asn(asn_file):
             self.get_data(self.test_dir, file)
 
-        pipe = Spec2Pipeline()
-        pipe.save_bsub = True
-        pipe.save_results = True
-        pipe.resample_spec.save_results = True
-        pipe.extract_1d.save_results = True
-        pipe.run(asn_file)
+        # Run the step
+        collect_pipeline_cfgs('cfgs')
+        Spec2Pipeline.call(asn_file, config_file='cfgs/calwebb_spec2.cfg', save_bsub=True)
 
+        # Test results.
         outputs = [('jw87600017001_02101_00002_nis_cal.fits',
                     'jw87600017001_02101_00002_nis_cal_ref.fits'),
                    ('jw87600017001_02101_00002_nis_x1d.fits',